Add base Credentials class
