+++ lib/gcloud/datastore/credentials.rb
@@ -12,8 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-require "json"
-require "signet/oauth_2/client"
+require "gcloud/credentials"
 
 module Gcloud
   module Datastore
@@ -23,28 +22,9 @@ module Gcloud
     # to the JSON keyfile downloaded from Google Cloud.
     #
     # https://developers.google.com/accounts/docs/application-default-credentials
-    class Credentials #:nodoc:
-      TOKEN_CREDENTIAL_URI = "https://accounts.google.com/o/oauth2/token"
-      AUDIENCE = "https://accounts.google.com/o/oauth2/token"
+    class Credentials < Gcloud::Credentials #:nodoc:
       SCOPE = ["https://www.googleapis.com/auth/datastore",
                "https://www.googleapis.com/auth/userinfo.email"]
-
-      ##
-      # The Signet client used to sign HTTP Requests.
-      attr_accessor :client #:nodoc:
-
-      # Create a new Credential object.
-      # This will raise if the JSON keyfile is incorrect.
-      def initialize keyfile
-        if keyfile.nil?
-          fail KeyfileError, "You must provide a keyfile to connect with."
-        elsif !File.exist?(keyfile)
-          fail KeyfileError, "The keyfile '#{keyfile}' is not a valid file."
-        end
-
-        options = JSON.parse(File.read(keyfile))
-        init_signet_client! options
-      end
 
       ##
       # Sign Oauth2 API calls.
@@ -56,26 +36,6 @@ module Gcloud
         request
       end
 
-      protected
-
-      ##
-      # Initialize the Signet client.
-      # Will raise if cannot acquire an access token.
-      def init_signet_client! options
-        client_opts = {
-          token_credential_uri: TOKEN_CREDENTIAL_URI,
-          audience: AUDIENCE,
-          scope: SCOPE,
-          issuer: options["client_email"],
-          signing_key: OpenSSL::PKey::RSA.new(options["private_key"])
-        }
-
-        @client = Signet::OAuth2::Client.new client_opts
-        @client.fetch_access_token!
-      end
-    end
-
-    class Credentials
       ##
       # Represent the empty credentials, useful for connecting
       # to a local devserver.
